% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_enrichment.R
\name{enrich_nordcan_cancer_case_dataset}
\alias{enrich_nordcan_cancer_case_dataset}
\title{NORDCAN Cancer Case Dataset}
\usage{
enrich_nordcan_cancer_case_dataset(
  x,
  iarccrgtools_exe_path,
  iarccrgtools_work_dir
)
}
\arguments{
\item{x}{\verb{[data.frame]} (mandatory, no default)
dataset of cancer cases}
}
\description{
Add new necessary columns to the cancer case dataset for use in computing
statistics.
}
\section{Format of input dataset}{

See https://github.com/CancerRegistryOfNorway/NORDCAN/wiki/Call-for-data---Incidence
}

\section{New columns}{

See https://github.com/CancerRegistryOfNorway/NORDCAN/wiki/Module-PreprocessEnrich
}

