% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_enrichment.R
\name{enrich_nordcan_cancer_record_dataset}
\alias{enrich_nordcan_cancer_record_dataset}
\title{NORDCAN Cancer Case Dataset}
\usage{
enrich_nordcan_cancer_record_dataset(
  x,
  iarccrgtools_exe_path,
  iarccrgtools_work_dir
)
}
\arguments{
\item{x}{\verb{[data.frame]} (mandatory, no default)
dataset of cancer cases}

\item{iarccrgtools_exe_path}{\verb{[character]} (mandatory, no default)
passed to \verb{[iarccrgtools::set_tools_exe_path]}}

\item{iarccrgtools_work_dir}{\verb{[character]} (mandatory, no default)
passed to \verb{[iarccrgtools::set_tools_work_dir]}}
}
\description{
Add new necessary columns to the cancer case dataset for use in computing
statistics.
}
\details{
\itemize{
\item enrich_nordcan_cancer_record_dataset takes a copy of \code{x} to ensure the
unprocessed dataset is not affected by the enrichment process.
\item surv_time is computed as the number of days
\code{end_of_followup - date_of_incidence}. Additionally,
when \code{autopsy == 1}, \code{surv_time} is set to zero.
}
}
\section{Format of input dataset}{

See https://github.com/CancerRegistryOfNorway/NORDCAN/wiki/Call-for-data---Incidence
}

\section{New columns}{

See https://github.com/CancerRegistryOfNorway/NORDCAN/wiki/Module-PreprocessEnrich
}

\section{Entity exceptions}{


Entities are added to the NORDCAN cancer record dataset in function
merge_nordcan_entity_columns. It first retrieves the table of ICD-10 to entity
conversion using \code{nordcancore::nordcan_metadata_icd10_to_entity()}.

The table of ICD-10 to entity definitions is merged with the cancer record
dataset by column "icd10".

There are a number of exceptions to the ICD-10 to entity conversion.

entity 300 is modfied by removing bascal cell carcinoma morpho-beh
combinations from it in function
nordcanpreprocessing:::merge_nordcan_entity_columns. the following
combinations are set to 888L in column entity_level_30:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{dt_basal <- data.table::CJ(
morpho = 8090:8098, beh = 3L
)
knitr::kable(dt_basal)
}\if{html}{\out{</div>}}\tabular{rr}{
   morpho \tab beh \cr
   8090 \tab 3 \cr
   8091 \tab 3 \cr
   8092 \tab 3 \cr
   8093 \tab 3 \cr
   8094 \tab 3 \cr
   8095 \tab 3 \cr
   8096 \tab 3 \cr
   8097 \tab 3 \cr
   8098 \tab 3 \cr
}


entity 319 is defined as all sub-types of 320 except 316 and 317. we define
it in function nordcanpreprocessing:::merge_nordcan_entity_columns
by first assigning 319 to every entity_level_30 value to 319 under
entity_level_20 == 320, and then defining 316 and 317 as exceptions to this.
entity 316 is defined as a sub-type of 320 based on morpho-beh combinations.
the combinations are as follows:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{dt_316 <-  data.table::CJ(
morpho = c(9530:9535, 9537:9539), beh = 0:3,
entity_level_20 = 320L
)
knitr::kable(dt_316)
}\if{html}{\out{</div>}}\tabular{rrr}{
   morpho \tab beh \tab entity_level_20 \cr
   9530 \tab 0 \tab 320 \cr
   9530 \tab 1 \tab 320 \cr
   9530 \tab 2 \tab 320 \cr
   9530 \tab 3 \tab 320 \cr
   9531 \tab 0 \tab 320 \cr
   9531 \tab 1 \tab 320 \cr
   9531 \tab 2 \tab 320 \cr
   9531 \tab 3 \tab 320 \cr
   9532 \tab 0 \tab 320 \cr
   9532 \tab 1 \tab 320 \cr
   9532 \tab 2 \tab 320 \cr
   9532 \tab 3 \tab 320 \cr
   9533 \tab 0 \tab 320 \cr
   9533 \tab 1 \tab 320 \cr
   9533 \tab 2 \tab 320 \cr
   9533 \tab 3 \tab 320 \cr
   9534 \tab 0 \tab 320 \cr
   9534 \tab 1 \tab 320 \cr
   9534 \tab 2 \tab 320 \cr
   9534 \tab 3 \tab 320 \cr
   9535 \tab 0 \tab 320 \cr
   9535 \tab 1 \tab 320 \cr
   9535 \tab 2 \tab 320 \cr
   9535 \tab 3 \tab 320 \cr
   9537 \tab 0 \tab 320 \cr
   9537 \tab 1 \tab 320 \cr
   9537 \tab 2 \tab 320 \cr
   9537 \tab 3 \tab 320 \cr
   9538 \tab 0 \tab 320 \cr
   9538 \tab 1 \tab 320 \cr
   9538 \tab 2 \tab 320 \cr
   9538 \tab 3 \tab 320 \cr
   9539 \tab 0 \tab 320 \cr
   9539 \tab 1 \tab 320 \cr
   9539 \tab 2 \tab 320 \cr
   9539 \tab 3 \tab 320 \cr
}


entity 317 is defined as a sub-type of 320 based on morpho-beh combinations
in function nordcanpreprocessing:::merge_nordcan_entity_columns.
the combinations are as follows:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{morpho_317 <- c(
9380, 9381, 9382, 9383, 9384, 9385, 9390, 9391, 9392, 9393,
9394, 9395, 9396, 9400, 9401, 9410, 9411, 9412, 9413, 9420, 9421,
9423, 9424, 9425, 9430, 9431, 9432, 9440, 9441, 9442, 9444, 9445,
9450, 9451, 9460, 9470, 9471, 9472, 9473, 9474, 9475, 9476, 9477,
9478, 9480
)
dt_317 <- data.table::CJ(morpho = morpho_317, beh = 1:3)
data.table::set(dt_317, j = "entity_level_20", value = 320L)
knitr::kable(dt_317)
}\if{html}{\out{</div>}}\tabular{rrr}{
   morpho \tab beh \tab entity_level_20 \cr
   9380 \tab 1 \tab 320 \cr
   9380 \tab 2 \tab 320 \cr
   9380 \tab 3 \tab 320 \cr
   9381 \tab 1 \tab 320 \cr
   9381 \tab 2 \tab 320 \cr
   9381 \tab 3 \tab 320 \cr
   9382 \tab 1 \tab 320 \cr
   9382 \tab 2 \tab 320 \cr
   9382 \tab 3 \tab 320 \cr
   9383 \tab 1 \tab 320 \cr
   9383 \tab 2 \tab 320 \cr
   9383 \tab 3 \tab 320 \cr
   9384 \tab 1 \tab 320 \cr
   9384 \tab 2 \tab 320 \cr
   9384 \tab 3 \tab 320 \cr
   9385 \tab 1 \tab 320 \cr
   9385 \tab 2 \tab 320 \cr
   9385 \tab 3 \tab 320 \cr
   9390 \tab 1 \tab 320 \cr
   9390 \tab 2 \tab 320 \cr
   9390 \tab 3 \tab 320 \cr
   9391 \tab 1 \tab 320 \cr
   9391 \tab 2 \tab 320 \cr
   9391 \tab 3 \tab 320 \cr
   9392 \tab 1 \tab 320 \cr
   9392 \tab 2 \tab 320 \cr
   9392 \tab 3 \tab 320 \cr
   9393 \tab 1 \tab 320 \cr
   9393 \tab 2 \tab 320 \cr
   9393 \tab 3 \tab 320 \cr
   9394 \tab 1 \tab 320 \cr
   9394 \tab 2 \tab 320 \cr
   9394 \tab 3 \tab 320 \cr
   9395 \tab 1 \tab 320 \cr
   9395 \tab 2 \tab 320 \cr
   9395 \tab 3 \tab 320 \cr
   9396 \tab 1 \tab 320 \cr
   9396 \tab 2 \tab 320 \cr
   9396 \tab 3 \tab 320 \cr
   9400 \tab 1 \tab 320 \cr
   9400 \tab 2 \tab 320 \cr
   9400 \tab 3 \tab 320 \cr
   9401 \tab 1 \tab 320 \cr
   9401 \tab 2 \tab 320 \cr
   9401 \tab 3 \tab 320 \cr
   9410 \tab 1 \tab 320 \cr
   9410 \tab 2 \tab 320 \cr
   9410 \tab 3 \tab 320 \cr
   9411 \tab 1 \tab 320 \cr
   9411 \tab 2 \tab 320 \cr
   9411 \tab 3 \tab 320 \cr
   9412 \tab 1 \tab 320 \cr
   9412 \tab 2 \tab 320 \cr
   9412 \tab 3 \tab 320 \cr
   9413 \tab 1 \tab 320 \cr
   9413 \tab 2 \tab 320 \cr
   9413 \tab 3 \tab 320 \cr
   9420 \tab 1 \tab 320 \cr
   9420 \tab 2 \tab 320 \cr
   9420 \tab 3 \tab 320 \cr
   9421 \tab 1 \tab 320 \cr
   9421 \tab 2 \tab 320 \cr
   9421 \tab 3 \tab 320 \cr
   9423 \tab 1 \tab 320 \cr
   9423 \tab 2 \tab 320 \cr
   9423 \tab 3 \tab 320 \cr
   9424 \tab 1 \tab 320 \cr
   9424 \tab 2 \tab 320 \cr
   9424 \tab 3 \tab 320 \cr
   9425 \tab 1 \tab 320 \cr
   9425 \tab 2 \tab 320 \cr
   9425 \tab 3 \tab 320 \cr
   9430 \tab 1 \tab 320 \cr
   9430 \tab 2 \tab 320 \cr
   9430 \tab 3 \tab 320 \cr
   9431 \tab 1 \tab 320 \cr
   9431 \tab 2 \tab 320 \cr
   9431 \tab 3 \tab 320 \cr
   9432 \tab 1 \tab 320 \cr
   9432 \tab 2 \tab 320 \cr
   9432 \tab 3 \tab 320 \cr
   9440 \tab 1 \tab 320 \cr
   9440 \tab 2 \tab 320 \cr
   9440 \tab 3 \tab 320 \cr
   9441 \tab 1 \tab 320 \cr
   9441 \tab 2 \tab 320 \cr
   9441 \tab 3 \tab 320 \cr
   9442 \tab 1 \tab 320 \cr
   9442 \tab 2 \tab 320 \cr
   9442 \tab 3 \tab 320 \cr
   9444 \tab 1 \tab 320 \cr
   9444 \tab 2 \tab 320 \cr
   9444 \tab 3 \tab 320 \cr
   9445 \tab 1 \tab 320 \cr
   9445 \tab 2 \tab 320 \cr
   9445 \tab 3 \tab 320 \cr
   9450 \tab 1 \tab 320 \cr
   9450 \tab 2 \tab 320 \cr
   9450 \tab 3 \tab 320 \cr
   9451 \tab 1 \tab 320 \cr
   9451 \tab 2 \tab 320 \cr
   9451 \tab 3 \tab 320 \cr
   9460 \tab 1 \tab 320 \cr
   9460 \tab 2 \tab 320 \cr
   9460 \tab 3 \tab 320 \cr
   9470 \tab 1 \tab 320 \cr
   9470 \tab 2 \tab 320 \cr
   9470 \tab 3 \tab 320 \cr
   9471 \tab 1 \tab 320 \cr
   9471 \tab 2 \tab 320 \cr
   9471 \tab 3 \tab 320 \cr
   9472 \tab 1 \tab 320 \cr
   9472 \tab 2 \tab 320 \cr
   9472 \tab 3 \tab 320 \cr
   9473 \tab 1 \tab 320 \cr
   9473 \tab 2 \tab 320 \cr
   9473 \tab 3 \tab 320 \cr
   9474 \tab 1 \tab 320 \cr
   9474 \tab 2 \tab 320 \cr
   9474 \tab 3 \tab 320 \cr
   9475 \tab 1 \tab 320 \cr
   9475 \tab 2 \tab 320 \cr
   9475 \tab 3 \tab 320 \cr
   9476 \tab 1 \tab 320 \cr
   9476 \tab 2 \tab 320 \cr
   9476 \tab 3 \tab 320 \cr
   9477 \tab 1 \tab 320 \cr
   9477 \tab 2 \tab 320 \cr
   9477 \tab 3 \tab 320 \cr
   9478 \tab 1 \tab 320 \cr
   9478 \tab 2 \tab 320 \cr
   9478 \tab 3 \tab 320 \cr
   9480 \tab 1 \tab 320 \cr
   9480 \tab 2 \tab 320 \cr
   9480 \tab 3 \tab 320 \cr
}


bladder/urinary tumours are only allowed to be in the following
morpho-beh combinations (and set to 999L otherwise):

\if{html}{\out{<div class="sourceCode r">}}\preformatted{dt_280 <- data.table::CJ(
morpho = c(8010L, 8120L, 8130L), beh = 0:2
)
knitr::kable(dt_280)
}\if{html}{\out{</div>}}\tabular{rr}{
   morpho \tab beh \cr
   8010 \tab 0 \cr
   8010 \tab 1 \cr
   8010 \tab 2 \cr
   8120 \tab 0 \cr
   8120 \tab 1 \cr
   8120 \tab 2 \cr
   8130 \tab 0 \cr
   8130 \tab 1 \cr
   8130 \tab 2 \cr
}


see function nordcanpreprocessing:::merge_nordcan_entity_columns for the
implementation.

the NORDCAN 8.2 stata code for this was

\if{html}{\out{<div class="sourceCode">}}\preformatted{`* limit included bladder/urinary tumors to 8010/2, 8120/1, 8120/2, 8130/1, 8130/2`
`replace entity_level1 = "999" if entity_level1 == "280" & inrange(beh, 0,2) & !inlist(morpho, 8010, 8120, 8130)`
}\if{html}{\out{</div>}}

we enforce all other entity_level_* columns to NA other than
entity_level_30 when entity_level_30 is either 888 or 999. see function
nordcanpreprocessing:::merge_nordcan_entity_columns for the implementation.
}

