% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_enrichment.R
\name{iarccrgtools}
\alias{iarccrgtools}
\alias{iarccrgtools_dataset}
\alias{iarccrgtools_tool}
\title{IARC CRG Tools}
\usage{
iarccrgtools_dataset(x, tool_name)

iarccrgtools_tool(x, tool_name, iarccrgtools_exe_path, iarccrgtools_work_dir)
}
\arguments{
\item{x}{\verb{[data.table]} (mandatory, no default)

the NORDCAN cancer record dataset}

\item{tool_name}{\verb{[character]} (mandatory, no default)

name of tool in IARC CRG Tools to use; one of the output of
\verb{[nordcancore::nordcan_iarccrgtools_tool_names()]}}

\item{iarccrgtools_exe_path}{\verb{[character]} (mandatory, no default)

path to executable of IARC CRG Tools; passed to
\link[iarccrgtools:set_tools_exe_path]{iarccrgtools::set_tools_exe_path}}

\item{iarccrgtools_work_dir}{\verb{[character]} (mandatory, no default)

path to a directory where inputs and outputs to IARC CRG Tools will be
stored; passed to \code{\link[iarccrgtools:set_tools_work_dir]{iarccrgtools::set_tools_work_dir()}}}
}
\description{
Pass NORDCAN cancer record dataset to IARC CRG Tools from R.
}
\details{
\itemize{
\item \code{iarccrgtools_dataset} collects a dataset from \code{x} that corresponds to the
requirements of the appropriate tool given in \code{tool_name};
see \verb{[iarccrgtools::create_example]} for the formats of each column, and
use \verb{[nordcancore::nordcan_metadata_column_name_set]} to see which columns are used
}

\itemize{
\item \code{iarccrgtools_tool} runs first \verb{[iarccrgtools::interact_with_tool]} and
then \verb{[iarccrgtools::connect_tool_results_to_observations]}; see those
functions for more information
}
}
