% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processed_cancer_death_count_dataset.R
\name{nordcan_processed_cancer_death_count_dataset}
\alias{nordcan_processed_cancer_death_count_dataset}
\title{NORDCAN Cancer Death Count Dataset}
\usage{
nordcan_processed_cancer_death_count_dataset(x)
}
\arguments{
\item{x}{\verb{[data.table]} (mandatory, no default)

the dataset of death counts as per the call for data}
}
\description{
Create the NORDCAN cancer death count dataset from raw data.
}
\examples{
dt <- data.table::data.table(
  year = 2018L,
  sex = 1L,
  region = 11L,
  agegroup = 1L,
  icd_code = "1751",
  icd_version = 7L,
  cancer_death_count = 10L
)

nordcancore::set_global_nordcan_settings(
  work_dir = ".",
  participant_name = "Norway",
  first_year_incidence = 1960L,
  first_year_mortality = 1980L,
  first_year_region = 1990L,
  last_year_incidence = 2020L,
  last_year_mortality = 2020L,
  last_year_survival = 2020L
)
cdcd <- nordcan_processed_cancer_death_count_dataset(x = dt)
}
