% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processed_cancer_record_dataset.R
\name{nordcan_processed_cancer_record_dataset}
\alias{nordcan_processed_cancer_record_dataset}
\title{NORDCAN Processed Cancer Record Dataset}
\usage{
nordcan_processed_cancer_record_dataset(x, iarccrgtools_exe_path)
}
\arguments{
\item{x}{\verb{[data.table]} (mandatory, no default)

the NORCDAN cancer record dataset before processing (as documented in
call for data)}

\item{iarccrgtools_exe_path}{\verb{[character]} (mandatory, no default)

passed to \link{iarccrgtools_tool}}
}
\description{
Form the NORDCAN processed cancer record dataset.
}
