% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_checks.R
\name{reports_assertions_tests}
\alias{reports_assertions_tests}
\alias{report_dataset_is_valid}
\alias{assert_dataset_is_valid}
\alias{assert_processed_cancer_record_dataset_is_valid}
\alias{assert_unprocessed_cancer_record_dataset_is_valid}
\alias{assert_general_population_size_dataset_is_valid}
\alias{assert_general_population_death_count_dataset_is_valid}
\alias{assert_national_population_life_table_is_valid}
\alias{test_dataset_is_valid}
\alias{report_dataset_columns_are_valid}
\alias{assert_dataset_columns_are_valid}
\alias{test_dataset_columns_are_valid}
\alias{report_column_is_valid}
\alias{assert_column_is_valid}
\alias{test_column_is_valid}
\title{Reports, Assertions, and Tests on NORDCAN Datasets and Columns}
\usage{
report_dataset_is_valid(x, dataset_name)

assert_dataset_is_valid(x, dataset_name)

assert_processed_cancer_record_dataset_is_valid(x)

assert_unprocessed_cancer_record_dataset_is_valid(x)

assert_general_population_size_dataset_is_valid(x)

assert_general_population_death_count_dataset_is_valid(x)

assert_national_population_life_table_is_valid(x)

test_dataset_is_valid(x, dataset_name)

report_dataset_columns_are_valid(x, column_names)

assert_dataset_columns_are_valid(x, column_names)

test_dataset_columns_are_valid(x, column_names)

report_column_is_valid(x, column_name)

assert_column_is_valid(x, column_name)

test_column_is_valid(x, column_name)
}
\arguments{
\item{x}{\verb{[data.frame]} (mandatory, no default)

dataset to report / assert / test on}

\item{dataset_name}{\verb{[character]} (mandatory, no default)

name of a NORDCAN dataset; must be one of the elements of output of
\code{nordcancore::nordcan_metadata_dataset_names()}; therefore see also
\verb{[nordcancore::nordcan_metadata_dataset_names]}}

\item{column_names}{\verb{[character]} (mandatory, no default)

names of columns to check}

\item{column_name}{\verb{[character]} (mandatory, no default)

name of column to check}
}
\description{
Functions to create reports, assert, and test NORDCAN datasets and columns
using functionality from package \code{dbc}.
}
\details{
\code{report_} functions produce a \code{data.frame} describing whether tests passed,
and if not, how they failed. See e.g. \verb{[dbc::tests_to_report]}.

\code{assert_} functions raise an error if any test in the corresponding report
did not pass. See e.g. \verb{[dbc::report_to_assertion]}.

\code{test_} functions return \code{TRUE} if all tests in the corresponding report
passed.
}
